Reorganize some vue templates.
