+++ .gitignore
@@ -5,6 +5,7 @@
 /public/js
 /public/mix-manifest.json
 /storage/*.key
+/bootstrap/cache/*.php
 /vendor
 .env
 .phpunit.result.cache
