+++ resources/js/Pages/Reports/Artwork/Index.vue
