+++ resources/js/Partials/Reports/IndividualReport.vue
